modularize data faker
